Refactor moves into their own component
